Move formatting to the "format phase" of the transport.
With the updated `winston-transport`, the transport write pipeline was changed, the place where the `log` method is called is different. Before, errors inside the log were always handled synchronously right after the write, but now in practice it's handled in `process.nextTick`.

This still runs in the same tick of the event loop, but after the current operation is finished. As a side effect, this means that errors thrown there are not caught in the current operation's context and try/catch blocks wrapping it won't work.

This means that in case of the format error which is tested in the broken unit test, no error is caught in the test so it fails. In fact, the logger does throw the error, but it's thrown only after the test finishes as an unhandled exception, which is very bas and I guess that's why this test exists.

In order to fix the issue, I moved the message formatting into the `format` part of the winston-transport pipeline, instead of inside of the `log` function itself. When there is a formatting error inside this pipeline, winston catches it on it's own, handles it internally and rethrows it, so the log action would throw it and it can be caught.